% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_decomp.R
\name{gnom_cor_decomp}
\alias{gnom_cor_decomp}
\title{Correlation decomposition of two signals measured along multiple chromosomes}
\usage{
gnom_cor_decomp(
  data,
  signals,
  chromosome,
  method = "pearson",
  rm.boundary = TRUE
)
}
\arguments{
\item{data}{data.table containing columns for variables to be correlated and a column with chromosome id if applicable.
data.table should be keyed by chromosome and then position. It is assumed that all positions in the data.table are equally spaced along the chromosome.
If not, results will not be valid. To get measurements at equally-spaced locations along chromosomes, you can interpolate your signals using the R function \code{approx()}.}

\item{signals}{character vector, names of columns to be correlated}

\item{chromosome}{character string, name of column containing chromosome id. To calculate results per chromosome, use chromosome = NA, and run function separately per chromosome.}

\item{method}{one of 'pearson' or 'spearman', indicating the type of correlation to calculate from wavelet coefficients. Note 'pearson' wavelet correlations for detail coefficients
do not strictly conform to the definition of the Pearson correlation as the product of means is not subtracted, but this gives an unbiased estimate of the wavelet correlation.
For scaling coefficients, the usual Pearson formula is used. Note also that the exact decomposition of the total correlation as described in manuscript is only valid using the
'pearson' wavelet and scaling correlations. \cr}

\item{rm.boundary}{logical, whether to remove boundary coefficients in the calculations. Default is FALSE.}
}
\value{
data.table containing:\tabular{ll}{
\code{level} \tab The level of the wavelet decomposition. Levels "d?" correspond to wavelet coefficients.
Levels "s?" correspond to scaling coefficients. Note 'pearson' correlations for wavelet coefficients don't subtract off the product of means.
Otherwise 'pearson' correlations for scaling coefficients or at the chromosome-level, and all 'spearman' correlations are calculated in the usual way.
Chromosome-level correlations are weighted by chromosome length.\cr
\tab \cr
\code{cor_n} \tab Correlation computed at the given level \cr
\tab \cr
\code{cor_jack} \tab Bias-corrected correlation from weighted block jackknife procedure. Details of procedure found at # https://reich.hms.harvard.edu/sites/reich.hms.harvard.edu/files/inline-files/wjack.pdf \cr
\tab \cr
\code{cor_jack_se} \tab Standard error of correlation computed from a weighted block jackknife across chromosomes. Only returned if number of chromosomes containing the given level is at least 6, otherwise returns NA. Currently only implemented for pearson correlations. \cr
}
}
\description{
Correlation decomposition of two signals measured along multiple chromosomes
}
\examples{
# Toy example - two white-noise signals "x" and "y" measured
# along six 'chromosomes'.

library(data.table)
data <- data.table(group = c(rep(1,1000), rep(2,1000), rep(3,750), rep(4,500), rep(5,500), rep(6,250)), x = rnorm(4000))
data[, y := x + rnorm(4000, mean=0, sd=2)]
signals <- c("x", "y"); chromosome <- "group"
gnom_cor_decomp(data, signals=signals, chromosome = chromosome)

}
