% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_decomp.R
\name{gnom_var_decomp}
\alias{gnom_var_decomp}
\title{Genome-wide variance decomposition}
\usage{
gnom_var_decomp(
  data,
  chromosome,
  signals,
  rm.boundary = TRUE,
  avg.over.chroms = TRUE
)
}
\arguments{
\item{data}{data.frame containing columns for chromosome id and signal measurements}

\item{chromosome}{character string, name of column containing chromosome id}

\item{signals}{character vector, names of columns containing the signals for which
variance decomposition is desired}

\item{rm.boundary}{logical, whether to remove boundary coefficients}

\item{avg.over.chroms}{logical, if FALSE, returns variance decomposition per chromosome
if true, variances are averaged over chromosomes}
}
\value{
if avg.over.chroms = TRUE, returns a data.table containing:\tabular{ll}{
\code{level} \tab The level of the wavelet decomposition. Values "dX"
correspond to wavelet coefficients, and corresponding variances are associated
with changes in localized average signal values over scale 2^(X-1). Values "sX" correspond to
scaling coefficients, and corresponding variances are associated with localized averages
on a scale of 2^X.  \cr
\tab \cr
\code{variance.X} \tab A weighted average across chromosomes of the magnitude of the
wavelet variance of signal "X" at the specified level of the decomposition.
Chromosomes without any coefficients at the specified level do not contribute to this average.
For the level of the entire chromosome, the value is a weighted variance of chromosome signal
means.
\cr
\tab \cr
\code{propvar.X} \tab A weighted average across chromosomes of the proportion of the
total variance of signal "X" contributed by the specified level of the decomposition.
Values are then adjusted to account for the total proportion of genomic-wide variance contributed
by differences in chromosome means. Thus, this has the interpretation of the total proportion
of genome-wide variance of "X" contributed by the specified level.
}
\description{
Performs a scale-based decomposition of the variance in signal measurements along a genome
containing multiple chromosomes. Wavelet variances are obtained using the Maximal Overlap
Discrete Wavelet Transform with Haar wavelets. By default boundary coefficients are
removed, which provides an unbiased estimate of the wavelet variance. A weighted average
of wavelet variances is taken over chromosmes, and chromosome-level variance is also
calculated.
}
\examples{
# Simulated example - two white-noise signals "x" and "y" measured along two chromosomes.
# Chromosomes differ in their means by the same amount, and signal y has larger variance than x.
library(data.table)
x <- data.table(x = c(rep(0, 30), rep(10, 70)) + rnorm(100),
                y = c(rep(0, 30), rep(10, 70)) + rnorm(100, 0, 10),
                group = c(rep(1, 30), rep(2, 70))
                position.id = c(1:30, 1:70))
plot(x[, c(x,y)])

wv <- gnom_wav_var(x, chromosome = "group", signals = c("x","y"))
with(wv, plot(variance.y ~ level))
with(wv, plot(propvar.y ~ level))

}
