% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_decomp.R
\name{modwt_lm_rsqrd}
\alias{modwt_lm_rsqrd}
\title{R squared from linear model of wavelet coefficients}
\usage{
modwt_lm_rsqrd(data, yvar, xvars, chromosome, rm.boundary = FALSE)
}
\arguments{
\item{data}{data.frame containing columns for chromosome id and signal measurements.
It is assumed that the rows contain datapoints that are evenly spaced along the sequence.
If not, you will need to interpolate your signal to a grid of evenly spaced points.}

\item{yvar}{character string, the response variable in the linear model}

\item{xvars}{character vector, the predictor variables in the linear model. Assumed no interaction}

\item{chromosome}{character string, the name of the column containing chromosome id}

\item{rm.boundary}{logical, whether to remove boundary coefficients (default FALSE)}
}
\value{
if avg.over.chroms = TRUE, returns a data.table containing:\tabular{ll}{
\code{level} \tab The level of the wavelet decomposition. Values "dX"
correspond to wavelet coefficients, and corresponding variances are associated
with changes in localized average signal values over scale 2^(X-1). Values "sX" correspond to
scaling coefficients, and corresponding variances are associated with localized averages
on a scale of 2^X.  \cr
\tab \cr
\code{rsqrd_n} \tab R squared from a linear model using the specified formula.
For wavelet coefficients, the regression is forced through the origin.
For scaling coefficients and chromosome means, an intercept is fit.
For the chromosome level, the regression of means is weighted by chromosome length.
\cr
\tab \cr
\code{rsqrd_jack} \tab The bias-corrected estimate of r squared from a weighted block jacknife with chromosomes as blocks.
Not reported if the given level contains < 4 blocks.
\cr
\tab \cr
\code{rsqrd_jack_se} \tab Standard error of the jacknife estimate.
}
}
\description{
Estimates r squared from a linear model of wavelet coefficients.
If separate chromosomes are present, also estimates r squared from linear model of chromosome means,
weighting by chromosome length. Weighted block jacknife with chromosomes as blocks
is performed to return bias-corrected estimates and standard errors of the bias-corrected estimate.
}
\examples{
# sample data with 20 chroms, where z is correlated with both x and y separately

library(data.table)

chrlens <- sample(x = c(100, 200, 300, 400, 500), size = 20, replace=TRUE)
chr_id <- NULL
for(i in 1:20){
  chr_id <- c(chr_id, rep(i, chrlens[i]))
}
data <- data.table(chrom = chr_id, x = rnorm(length(chr_id)), y = rnorm(length(chr_id)))
data[, z := 4*x - 2*y + rnorm(length(chr_id), sd = 5)]

modwt_lm_rsqrd(data=data, yvar = "z", xvars = c("x", "y"), chromosome = "chrom")

}
