% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_modwts.R
\name{multi_modwt_1var}
\alias{multi_modwt_1var}
\title{modwt of a signal on one chromosome of a genome with multiple chromosomes}
\usage{
multi_modwt_1var(u, all_cols, rm.boundary = TRUE)
}
\arguments{
\item{u}{a vector of signal measurements}
}
\value{
a data.table with 2 columns, level of the decomposition,
coefficient values
}
\description{
Helper function for dijoint_modwts below.
Wrapper for modwt function from waveslim. Sets coefficients to NA for levels
that are not present on the focal chromosome but that are present on others.
}
