% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_modwts.R
\name{multi_modwts}
\alias{multi_modwts}
\title{Maximal Overlap Discrete Wavelet Transform of signals measured along
a set of disjoint sequences.}
\usage{
multi_modwts(data, chromosome, signals, rm.boundary = F)
}
\arguments{
\item{data}{a data.frame or data.table containing a column for chromosome ID and column(s) with signal values.}

\item{chromosome}{character string, the name of the column with chromosome ID}

\item{signals}{character vector, names of the columns with signal values}
}
\value{
data.table containing:\tabular{ll}{
\code{level} \tab The level of the wavelet decomposition. Levels "d?" correspond to wavelet coefficients.
Levels "s?" correspond to scaling coefficients. \cr
\tab \cr
\code{coefficient.X} \tab Coefficient values for the signal named "X". \cr
\tab \cr
\code{position.id} \tab  End position of the wavelet or scaling filter that produced this coefficient.
As boundary coefficients are removed, the first position.id is the end position of the wavelet that starts
at the beginning of the sequence.  \cr
}
}
\description{
Wrapper function for functions from package \code{waveslim}.
Returns a table of coefficients from the Maximal Overlap Discrete Wavelet Transform (MODWT)
performed on multiple signal measurements (i.e. ancestry state and recombination rate)
along multiple disjoint sequences (i.e. chromosomes). By default the 'brick wall' boundary coefficient
is applied, meaning that boundary coefficients are removed from the output.
}
\examples{
tbd

}
