% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theory.R
\name{wavelet_variance_equilibrium}
\alias{wavelet_variance_equilibrium}
\title{Calculate expected wavelet variance of ancestry state under neutral admixture pulse model with constant population size}
\usage{
wavelet_variance_equilibrium(n.pop, n.sample, unit.dist, level, gen, alpha)
}
\arguments{
\item{n.pop}{Numeric vector. Total number of chromosomes in the population. e.g. should be 2N for a diploid population}

\item{n.sample}{numeric. returns wavelet variance for mean ancestry in a sample of this size. \code{n.sample} of 1
gives expectation for a single chromosome}

\item{unit.dist}{Unit distance, i.e. spacing between adjacent hypothetical loci, in Morgans}

\item{level}{Integer vector giving levels at which expectation will be returned.
Level 1 would correspond to expected variance associated with changes occurring over the unit distance.}

\item{gen}{Integer vector of generations for which expectation should be returned.
Note that generations correspond to number of generations of recombination since the admixture pulse,
i.e. number of meioses starting with meiosis in F1s.}

\item{alpha}{numeric. the admixture proportion.}
}
\value{
a data.table with columns corresponding to the arguments given and column \code{variance}
giving the expectations.
}
\description{
Calculate expected wavelet variance of ancestry state under neutral admixture pulse model with constant population size
}
\examples{
wavelet_variance_equilibrium(n.pop = 2000, n.sample=1, unit.dist = 2^-10, level = 1:10, gen = c(100,1000), alpha = 0.5)
}
