% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theory.R
\name{wavelet_variance_F2}
\alias{wavelet_variance_F2}
\title{Calculate expected wavelet variance of ancestry state for F2 generation}
\usage{
wavelet_variance_F2(n.pop, n.sample, unit.dist, level, alpha)
}
\arguments{
\item{n.pop}{Numeric vector. Total number of chromosomes in the population. e.g. should be 2N for a diploid population}

\item{n.sample}{numeric. returns wavelet variance for mean ancestry in a sample of this size. \code{n.sample} of 1
gives expectation for a single chromosome}

\item{unit.dist}{Unit distance, i.e. spacing between adjacent hypothetical loci, in Morgans}

\item{level}{Integer vector giving levels at which expectation will be returned.
Level 1 would correspond to expected variance associated with changes occurring over the unit distance.}

\item{alpha}{numeric. the admixture proportion.}
}
\value{
a data.table with columns corresponding to the arguments given and column \code{variance}
giving the expectations.
}
\description{
Calculate expected wavelet variance of ancestry state for F2 generation
}
