% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_decomp.R
\name{gnom_var_decomp}
\alias{gnom_var_decomp}
\title{Genome-wide variance decomposition}
\usage{
gnom_var_decomp(
  data,
  chromosome,
  signals,
  rm.boundary = FALSE,
  avg.over.chroms = TRUE
)
}
\arguments{
\item{data}{data.frame containing columns for chromosome id and signal measurements.
It is assumed that the rows contain datapoints that are evenly spaced along the sequence.
If not, you will need to interpolate your signal to a grid of evenly spaced points.}

\item{chromosome}{character string, name of column containing chromosome id}

\item{signals}{character vector, names of columns containing the signals for which
variance decomposition is desired}

\item{rm.boundary}{logical, whether to remove boundary coefficients}

\item{avg.over.chroms}{logical, if FALSE, returns variance decomposition per chromosome
if true, variances are averaged over chromosomes}
}
\value{
if avg.over.chroms = TRUE, returns a data.table containing:\tabular{ll}{
\code{level} \tab The level of the wavelet decomposition. Values "dX"
correspond to wavelet coefficients, and corresponding variances are associated
with changes in localized average signal values over scale 2^(X-1). Values "sX" correspond to
scaling coefficients, and corresponding variances are associated with localized averages
on a scale of 2^X.  \cr
\tab \cr
\code{variance.X} \tab A weighted average across chromosomes of the magnitude of the
wavelet variance of signal "X" at the specified level of the decomposition.
Chromosomes without any coefficients at the specified level do not contribute to this average.
For the level of the entire chromosome, the value is a weighted variance of chromosome signal
means.
\cr
\tab \cr
\code{propvar.X} \tab A weighted average across chromosomes of the proportion of the
total variance of signal "X" contributed by the specified level of the decomposition.
Values are then adjusted to account for the total proportion of genomic-wide variance contributed
by differences in chromosome means. Thus, this has the interpretation of the total proportion
of genome-wide variance of "X" contributed by the specified level.
This interpretation depends on having a high quality chromosome-level genome assembly.
\cr
\tab \cr
\code{variance.X.jack} \tab Weighted jacknife estimate of the variance. At each level of the decomposition,
only chromosomes containing coefficients at that level are used in the jacknife.
\cr
\tab \cr
\code{variance.X.jack.se} \tab Standard error of the jacknife estimate.
}
}
\description{
Performs a scale-based decomposition of the variance in signal measurements along a genome
containing multiple chromosomes. Wavelet variances are obtained using the Maximal Overlap
Discrete Wavelet Transform with Haar wavelets. By default boundary coefficients are not
removed, but these can be ignored in the variance estimates. A weighted average
of wavelet variances is taken over chromosmes, and chromosome-level variance is also
calculated.
}
\examples{
# Simulated example - two white-noise signals "x" and "y" measured along two chromosomes. Signal y has larger variance than x.

library(data.table)
library(ggplot(2))
d <- data.table(chr = c(rep(1,1000), rep(2,1000), rep(3,750), rep(4,500), rep(5,500), rep(6,250)), x = rnorm(4000), y = rnorm(4000, sd=10))
gnom_var_decomp(d, chromosome = "chr", signals = c("x","y"))
}
